// src/types.ts
export type CastEventType = "o" | "i" | "r" | "m";
export type CastEvent = [time: number, type: CastEventType, data: string];